requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
crispy-tailwind==1.0.3
cryptography==44.0.3
dj-database-url==2.3.0
Django==5.2
django-admin-honeypot==1.1.0
django-allauth==65.7.0
django-axes==7.1.0
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-csp==4.0
django-filter==25.1
django-formtools==2.5.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-modeltranslation==0.19.14
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-ratelimit==4.1.0
django-recaptcha==4.1.0
django-secure==1.0.2
django-storages==1.14.6
django-two-factor-auth==1.17.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
qrcode==7.4.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
# Production requirements
sentry-sdk==2.0.0  # Error tracking
django-environ==0.11.2  # Better environment variable handling
